src/speckit_core/pyproject.toml
# =============================================================================
# Package: speckit-core (Shared Library)
# =============================================================================

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "speckit-core"
version = "0.1.0"
description = "Shared core library for SpecKit tools"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
